requirements.txt
openpyxl==3.1.5
pandas==2.2.3
plyer==2.1.0
pystray==0.19.5
Pillow==10.4.0
pywebview==5.1
